adjust to big int

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,8 +20,8 @@ class CreateUsersTable extends Migration
             $table->timestamp('date_of_birth')->nullable();
             $table->integer('last_degree_id')->unsigned()->nullable();
             $table->integer('major_id')->unsigned()->nullable();
-            $table->integer('expected_salary_lower')->nullable();
-            $table->integer('expected_salary_upper')->nullable();
+            $table->bigInteger('expected_salary_lower')->nullable();
+            $table->bigInteger('expected_salary_upper')->nullable();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
